requirements.txt
pywebview>=4.0
Flask>=2.3.0
